資金操作ログの実装

+++ structs/src/remining_money/mod.rs
@@ -5,10 +5,9 @@ use super::cash_history::{CashHistory, Badget};
 use super::day::Day;
 struct ReminingMoney{
     logs:BTreeMap<Day, Vec<CashHistory>>,
-    remining_money:HashMap<Badget,u64>,
+    remining_money:HashMap<Badget,i128>,
 }
 
 #[test]
 fn remining_money_test(){
-
 }
